renamed 'royal_render_module.py' to 'addon.py'
